+++ packages/plugins/Gfi/src/store/actions/setup.ts
@@ -4,8 +4,6 @@ import getCluster from '@polar/lib-get-cluster'
 import { getTooltip, Tooltip } from '@polar/lib-tooltip'
 import Overlay from 'ol/Overlay'
 import { Feature } from 'ol'
-import { DragBox, Draw, Modify } from 'ol/interaction'
-import { platformModifierKeyOnly } from 'ol/events/condition'
 import { GfiGetters, GfiState } from '../../types'
 import { getOriginalFeature } from '../../utils/getOriginalFeature'
 
@@ -23,46 +21,6 @@ export function setupCoreListener(
       (feature) => dispatch('setOlFeatureInformation', { feature }),
       { deep: true }
     )
-  }
-}
-
-export function setupMultiSelection({
-  dispatch,
-  getters,
-  rootGetters,
-}: PolarActionContext<GfiState, GfiGetters>) {
-  if (getters.gfiConfiguration.boxSelect) {
-    const dragBox = new DragBox({ condition: platformModifierKeyOnly })
-    dragBox.on('boxend', () =>
-      dispatch('getFeatureInfo', {
-        coordinateOrExtent: dragBox.getGeometry().getExtent(),
-        modifierPressed: true,
-      })
-    )
-    rootGetters.map.addInteraction(dragBox)
-  }
-  if (getters.gfiConfiguration.directSelect) {
-    rootGetters.map.on('click', ({ coordinate, originalEvent }) => {
-      const isDrawing = rootGetters.map
-        .getInteractions()
-        .getArray()
-        .some(
-          (interaction) =>
-            // these indicate other interactions are expected now
-            interaction instanceof Draw ||
-            interaction instanceof Modify ||
-            // @ts-expect-error | internal hack to detect it from @polar/plugin-draw
-            interaction._isDeleteSelect
-        )
-      if (!isDrawing) {
-        dispatch('getFeatureInfo', {
-          coordinateOrExtent: coordinate,
-          modifierPressed: navigator.userAgent.includes('Mac')
-            ? originalEvent.metaKey
-            : originalEvent.ctrlKey,
-        })
-      }
-    })
   }
 }