+++ packages/plugins/Pins/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## unpublished
+## 3.0.0
 
 - Breaking: Upgrade peerDependency `ol` from `^9.2.4` to `^10.3.1`.
 - Feature: Recognize the multi-select-`Draw` of `@polar/plugin-gfi`.